Refactor agents into separate modules
